+++ run_tests.py
@@ -11,7 +11,7 @@ Skip slow tests: ::
 
     python run_tests.py fast
     
-Code imported from textblob-fr sample extension.
+Code imported from ``textblob-fr`` sample extension.
 
 :repo: `https://github.com/sloria/textblob-fr`_
 :source: run_tests.py