+++ tests/test_sentiments.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env
 # -*- coding: utf-8 -*-
-'''Code imported from textblob-fr sample extension.
+'''Code imported from ``textblob-fr`` sample extension.
 
 :repo: `https://github.com/sloria/textblob-fr`_
 :source: tests/test_sentiments.py
@@ -12,18 +12,25 @@
 import unittest
 from nose.tools import *  # PEP8 asserts
 
-from textblob import TextBlob
+from textblob_de import TextBlobDE as TextBlob
 from textblob_de import PatternAnalyzer as DeAnalyzer
+from textblob_de.tokenizers import get_tokenizer, PatternTokenizer
 
 
 class TestPatternAnalyzer(unittest.TestCase):
 
     def setUp(self):
+        self.tokenizer = PatternTokenizer()
         self.analyzer = DeAnalyzer()
         self.neg1 = u"Das Auto ist schrecklich."
         self.pos1 = u"Das Auto ist schön."
         self.pos2 = u"Die Katze ist nicht böse."
         self.neg2 = u"Dieser Hund ist nicht nett."
+        
+        setattr(get_tokenizer, 'tokenizer', self.tokenizer)
+        
+    def tearDown(self):
+        delattr(get_tokenizer, 'tokenizer')          
 
     def test_analyze(self):
         pos_sentiment = self.analyzer.analyze(self.pos1)