+++ tests/test_taggers.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-'''Code imported from textblob-fr sample extension.
+'''Code imported from ``textblob-fr`` sample extension.
 
 :repo: `https://github.com/sloria/textblob-fr`_
 :source: tests/test_taggers.py
@@ -13,15 +13,22 @@ import unittest
 import logging
 from nose.tools import *  # PEP8 asserts
 
-from textblob import TextBlob
+from textblob_de import TextBlobDE as TextBlob
+from textblob_de.tokenizers import PatternTokenizer, NLTKPunktTokenizer, get_tokenizer
 from textblob_de import PatternTagger
 
 
-class TestPatternTagger(unittest.TestCase):
+class TestPatternTaggerWithNLTKTok(unittest.TestCase):
 
     def setUp(self):
+        self.tokenizer = NLTKPunktTokenizer()
         self.tagger = PatternTagger()
         self.text = u"Das ist ein schönes Auto"
+        
+        setattr(get_tokenizer, 'tokenizer', self.tokenizer)
+        
+    def tearDown(self):
+        delattr(get_tokenizer, 'tokenizer')        
 
     def test_tag(self):
         tags = self.tagger.tag(self.text)
@@ -31,7 +38,34 @@ class TestPatternTagger(unittest.TestCase):
             assert_equal(word_tag[0], words[i])
 
     def test_tag_blob(self):
-        blob = TextBlob(self.text, pos_tagger=self.tagger)
+        blob = TextBlob(self.text, pos_tagger=self.tagger, tokenizer=self.tokenizer)
+        tags = blob.tags
+        logging.debug("tags: {0}".format(tags))
+        words = self.text.split()
+        for i, word_tag in enumerate(tags):
+            assert_equal(word_tag[0], words[i])
+
+class TestPatternTaggerWithPatternTok(unittest.TestCase):
+
+    def setUp(self):
+        self.tokenizer = PatternTokenizer()
+        self.tagger = PatternTagger()
+        self.text = u"Das ist ein schönes Auto"
+        
+        setattr(get_tokenizer, 'tokenizer', self.tokenizer)
+        
+    def tearDown(self):
+        delattr(get_tokenizer, 'tokenizer')        
+
+    def test_tag(self):
+        tags = self.tagger.tag(self.text)
+        logging.debug("tags: {0}".format(tags))
+        words = self.text.split()
+        for i, word_tag in enumerate(tags):
+            assert_equal(word_tag[0], words[i])
+
+    def test_tag_blob(self):
+        blob = TextBlob(self.text, pos_tagger=self.tagger, tokenizer=self.tokenizer)
         tags = blob.tags
         logging.debug("tags: {0}".format(tags))
         words = self.text.split()