+++ textblob_de/__init__.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 from textblob_de.blob import TextBlobDE
+from textblob_de.tokenizers import NLTKPunktTokenizer, PatternTokenizer
 from textblob_de.taggers import PatternTagger
 from textblob_de.parsers import PatternParser
 from textblob_de.sentiments import PatternAnalyzer
 
-__version__ = '0.1.4a4'
+
+__version__ = '0.2.0a1'
 __author__ = 'Markus Killer'
 __license__ = "MIT"