+++ textblob_de/inflect.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-'''Code adapted from the pattern.de library.
+'''Code adapted from the ``pattern.de`` library.
 
 :repo: `https://github.com/clips/pattern`_
 :source: pattern/text/de/inflect.py