+++ textblob_de/taggers.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-'''Code adapted from textblob-fr sample extension.
+'''Code adapted from ``textblob-fr`` sample extension.
 
 :repo: `https://github.com/sloria/textblob-fr`_
 :source: textblob_fr/taggers.py
@@ -13,6 +13,8 @@ from __future__ import absolute_import
 from textblob.base import BaseTagger
 from textblob_de.de import tag as pattern_tag
 
+from textblob_de.tokenizers import get_tokenizer
+
 
 class PatternTagger(BaseTagger):
 
@@ -23,4 +25,6 @@ class PatternTagger(BaseTagger):
 
     def tag(self, sentence, tokenize=True):
         '''Tag a string `sentence`.'''
-        return pattern_tag(sentence, tokenize)
+        # for future implementations (needs to be changed in BaseBlob and BaseTagger)
+        #def tag(self, sentence, tokenizer, tokenize=True)
+        return pattern_tag(sentence, get_tokenizer(), tokenize)